fix lỗi vặt
thêm pagerouter ở login sang forgot, sang signup, bỏ mũi tên back ở profile, chỉnh màu checkout page
