+++ App/flower_store/lib/screens/store.main.screen.dart
@@ -4,7 +4,7 @@ import 'package:flower_store/constants/colors.dart';
 import 'package:flower_store/models/menu.model.dart';
 import 'package:flower_store/models/product.dart';
 import 'package:flower_store/screens/favorite.screen.dart';
-import 'package:flower_store/screens/main/profile.screen.dart';
+import 'package:flower_store/screens/user/profile.screen.dart';
 import 'package:flower_store/shared/components/custom_drawer.dart';
 import 'package:flower_store/shared/components/main_page_header.dart';
 import 'package:flower_store/shared/widget/carousel.dart';
@@ -105,7 +105,7 @@ class _StoreMainScreenState extends State<StoreMainScreen> {
                       ProductListview(
                         products: productsFavorite,
                         title: 'Your favorite',
-                        navigator: FavoriteScreen(),
+                        navigator: const FavoriteScreen(),
                         seeAll: true,
                       ),
                       ProductListview(
